backend/requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
pydantic

langchain-community
langchain-groq
faiss-cpu
sentence-transformers